smarter access for blocks/lines

+++ ScanLHA/runner.py
@@ -3,7 +3,6 @@ import logging
 from subprocess import Popen, STDOUT, PIPE
 from .slha import parseSLHA
 from random import randrange
-from numpy import nan
 import os
 
 class Runner():
@@ -26,6 +25,7 @@ class Runner():
         # TODO: copy binary into ram, execute from ram directory
 
     def run(self, params):
+        slha = {}
         # TODO: Too long filenames/argument for subprocess?
         # fname = '_'.join(['{}.{}'.format(p,v) for p,v in params.items()])
         fname = '%030x' % randrange(16**30)
@@ -37,27 +37,32 @@ class Runner():
                 params = defaultdict(str, { '%{}%'.format(p) : v for p,v in params.items() })
                 inputf.write(self.tpl.format_map(params))
             except KeyError:
-                logging.error("Could not substitute ", params)
-                return { 'log': 0 }
+                err = "Could not substitute {}.".format(params)
+                logging.error(err)
+                return { 'log': err }
 
         proc = Popen([self.config['binary'], fin, fout], stderr=STDOUT, stdout=PIPE)
         stdout, stderr = proc.communicate(timeout=self.timeout)
+        if self.config.get('remove_slha', True):
+            try:
+                os.remove(fin)
+            except FileNotFoundError:
+                logging.error('Input file {} missing?'.format(fin))
         if os.path.isfile(fout):
             slha = parseSLHA(fout, self.blocks)
-            if not self.config.get('keep_slha', False):
+            try:
                 os.remove(fout)
-                os.remove(fin)
-            return slha
-        if not self.config.get('keep_log', False):
-            return { 'log': nan }
-        if stdout or stderr:
-            log = 'parameters: %s\n' % params
-            with open(flog, 'w') as logf:
-                log += 'stdout: {}\nstderr: {}\n\n'.format(
-                            stdout.decode('utf8').strip(),
-                            stderr.decode('utf8').strip()
-                        )
-                logf.write(log)
-        if os.path.isfile(flog):
+            except Exception as e:
+                logging.error('Could not remove output file {}: {}'.format(fout, str(e)))
+
+        if (stdout or stderr) and self.config.get('keep_log', True):
+            stdout = stdout.decode('utf8').strip() if stdout else ' '
+            stderr = stderr.decode('utf8').strip() if stderr else ' '
+            log = 'parameters: {}\nstdout: {}\nstderr: {}\n\n'.format(params,stdout, stderr)
+            slha.update({ 'log': log })
+            if self.config.get('logfile', False):
+                with open(flog, 'w') as logf:
+                    logf.write(log)
+                slha.update({ 'log': flog })
             logging.debug(log)
-            return { 'log': flog }
+        return slha
